golbal context provider created for data passing
